Evaluate bristol circuits
Adds functionality to evaluate circuits in Bristol format. For this,
the structure returned from the parsed bristol file `circuit::Bristol`
is transformed into the internal `circuit::Circuit` representation.
Currently, this requires a rewriting of wire labels, which might be
problematic in the future when large circuits should be evaluated in a
streaming fashion, as the whole id mapping for the circuit needs to be
kept in (in memory? maybe it can be mmapped once it gets to big...).

Also adds a few tests, evaluating adders, AES and sha256. As those
circuits are a little bigger, evaluating the tests with --release might
be faster.