Add trusted MT Provider
Adds a MTProvider trait and two implementations of it.
The TrustedMTProvider can be used to delegate the sampling of MT's
to a trusted party. See the examples/trusted_party_mts.rs of how
to use the server. The sha256 example is changed to use a trusted
mt provider if specified or the insecure provider by default.