Fix: Fix a race condition in the circuit execution
Due to remoc sending messages in a background task, it can happen that
sent messages are lost if the process exits before the message is sent.
In this case, the other party misses the last layer for the execution
and panicks. This commit adds a synchronization message which both
parties sent. If a party does not receive the message, a warning is
logged, but the execution still completed correctly, as this bug only
cases the last message to potentially by lost (now the sync msg).